+++ CHANGES.md
@@ -19,6 +19,7 @@ Change Log
 * Ported a support for `GpxCatalogItem`.
 * Fixes a bug that showed the chart download button when there is no downloadable source.
 * Ensure the `chartDisclaimer` is passed from catalog items to derived chart items.
+* Only use an arcgis portal item extent to calculate `rectangle` when we have valid one.
 * [The next improvement]
 
 #### 8.0.0-alpha.53